Renamed CustomSelectField.tsx and added rightIcon to "First Nation Heritage" Field.
